Move fixed_val_tiles + fix data access
